Life View tool for remote debug

+++ MainForm.cs
@@ -1711,31 +1711,37 @@ namespace com.clusterrr.hakchi_gui
             {
                 if (WaitingClovershellForm.WaitForDevice(this))
                 {
-                    var screenshot = WorkerForm.TakeScreenshot();
-                    var screenshotPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName() + ".png");
-                    screenshot.Save(screenshotPath, ImageFormat.Png);
-                    var showProcess = new Process()
+                    if (Control.ModifierKeys != Keys.Shift)
                     {
-                        StartInfo = new ProcessStartInfo()
+                        var screenshot = WorkerForm.TakeScreenshot();
+                        var screenshotPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName() + ".png");
+                        screenshot.Save(screenshotPath, ImageFormat.Png);
+                        var showProcess = new Process()
                         {
-                            FileName = screenshotPath
-                        }
-                    };
-                    showProcess.Start();
-                    new Thread(delegate ()
+                            StartInfo = new ProcessStartInfo()
+                            {
+                                FileName = screenshotPath
+                            }
+                        };
+                        showProcess.Start();
+                        new Thread(delegate ()
+                        {
+                            try
+                            {
+                                Thread.Sleep(5000);
+                                showProcess.WaitForExit();
+                            }
+                            catch { }
+                            try
+                            {
+                                File.Delete(screenshotPath);
+                            }
+                            catch { }
+                        }).Start();
+                    } else // Shift pressed
                     {
-                        try
-                        {
-                            Thread.Sleep(5000);
-                            showProcess.WaitForExit();
-                        }
-                        catch { }
-                        try
-                        {
-                            File.Delete(screenshotPath);
-                        }
-                        catch { }
-                    }).Start();
+                        new LifeViewForm().ShowDialog();
+                    }
                 }
             }
             catch (Exception ex)
